perf: Changed couple variable to short and bytes for memory improvement in database
Changed variables in TaskDataModel (LevelOfTask) and UserDataModel (InviteCount and Level) to shorts and bytes as they cannot exceed the maximum values in real-life scenario.

+++ GrabDemSite/Interfaces/Interfaces.cs
@@ -23,9 +23,9 @@ namespace GrabDemSite.Interfaces
         double Balance { get; set; }
         double MoneySpent { get; set; }
         string WalletAddress { get; set; }
-        int InviteCount { get; set; }
+        short InviteCount { get; set; }
         double PlayMoney { get; set; }
-        int Level { get; set; }
+        byte Level { get; set; }
     }
     public interface IMethodsCall
     {
